requirements.txt
Flask
tensorflow-cpu
joblib
numpy
gunicorn
